refactor: made more stuff internal
